Materials/requirements.txt
numpy
opencv-python
faiss-cpu
dlib
fastapi
uvicorn
asyncio-mqtt
